+++ src/main/java/org/jahia/modules/htmlfiltering/impl/config/model/ElementCfg.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jahia.modules.htmlfiltering.configuration;
+package org.jahia.modules.htmlfiltering.impl.config.model;
 
 import java.util.List;
