+++ src/main/java/org/jahia/modules/htmlfiltering/impl/config/model/WorkspaceCfg.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jahia.modules.htmlfiltering.configuration;
+package org.jahia.modules.htmlfiltering.impl.config.model;
 
 import org.jahia.modules.htmlfiltering.interceptor.HtmlFilteringInterceptor;
 import org.jahia.services.content.JCRNodeWrapper;
