+++ tests/graphql.config.yml
@@ -1,0 +1,10 @@
+schema: schema.graphql
+extensions:
+  endpoints:
+    Default GraphQL Endpoint:
+      url: http://localhost:8080/modules/graphql
+      headers:
+        user-agent: JS GraphQL
+        Authorization: Basic cm9vdDpyb290MTIzNA==
+        Origin: http://localhost:8080
+      introspect: true
